DMD-93 - refactor snowflake helper methods to trait
